float

+++ Assets/Scripts/Network/PacketSend.cs
@@ -24,7 +24,7 @@ public class PacketSend
         {
             p.Write(pl.NetworkID);
             p.WriteUNICODE(TestRandomUnicode);
-            p.Write(DateTime.Now.Ticks);
+            p.Write((float)DateTime.Now.Ticks);
             return pl.SendPacket(p);
 
         };
@@ -117,7 +117,7 @@ public class PacketSend
         using (packet p = new packet((int)ClientPackets.Test_Packet))
         {
             p.WriteUNICODE(TestRandomUnicode);
-            p.Write(DateTime.Now.Ticks);
+            p.Write((float)DateTime.Now.Ticks);
             return SendToServer(p);
